% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{account}
\alias{account}
\title{Set account for a resource profile}
\usage{
account(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Account name}
}
\value{
Updated profile object
}
\description{
Set account for a resource profile
}
\note{
Use res_account() to avoid naming collisions in user code.
}
\examples{
\dontrun{
account(job)
}
}
