% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{artifact}
\alias{artifact}
\title{Alias for file_ref function}
\usage{
artifact()
}
\value{
File reference schema
}
\description{
Alias for file_ref function
}
\examples{
\dontrun{
artifact("model_fit", value = my_model, type = "rds")
}
}
