% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{as.list.parade_profile}
\alias{as.list.parade_profile}
\title{Convert profile to list for slurm_resources()}
\usage{
\method{as.list}{parade_profile}(x, ...)
}
\arguments{
\item{x}{A resource profile object}

\item{...}{Additional arguments (unused)}
}
\value{
List of resources
}
\description{
Convert profile to list for slurm_resources()
}
\examples{
\dontrun{
prof <- profile_get("standard")
as.list(prof)
}

}
