% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{as_function}
\alias{as_function}
\title{Convert formula to function (simplified version)}
\usage{
as_function(formula)
}
\arguments{
\item{formula}{Formula to convert}
}
\value{
Function
}
\description{
Convert formula to function (simplified version)
}
\keyword{internal}
