% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_results.R
\name{as_tibble.parade_jobset}
\alias{as_tibble.parade_jobset}
\title{Convert jobset to tibble}
\usage{
\method{as_tibble}{parade_jobset}(x, ...)
}
\arguments{
\item{x}{A parade_jobset object}

\item{...}{Additional arguments (unused)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with one row per job.
}
\description{
Convert jobset to tibble
}
\examples{
\dontrun{
jobs <- slurm_map(1:4, function(x) x^2)
tibble::as_tibble(jobs)
}

}
