% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{batch_resources}
\alias{batch_resources}
\title{Create a list of SLURM resources with friendly parsing.}
\usage{
batch_resources(
  partition = NULL,
  time = NULL,
  nodes = NULL,
  ntasks = NULL,
  ntasks_per_node = NULL,
  cpus_per_task = NULL,
  ncpus = NULL,
  mem = NULL,
  account = NULL,
  qos = NULL,
  modules = NULL,
  omp_num_threads = NULL
)
}
\arguments{
\item{partition, time, nodes, ntasks, ntasks_per_node, cpus_per_task, ncpus, mem, account, qos, modules, omp_num_threads}{Various resource fields.}
}
\value{
Named list suitable for batchtools' \code{submitJobs(resources=...)} and the template.
}
\description{
Create a list of SLURM resources with friendly parsing.
}
