% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{cancel}
\alias{cancel}
\title{Cancel all jobs in a jobset}
\usage{
cancel(x, ...)
}
\arguments{
\item{x}{A parade_jobset object}
}
\value{
The jobset (invisibly)
}
\description{
Cancel all jobs in a jobset
}
