% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{chr}
\alias{chr}
\title{Create a character type specification}
\usage{
chr()
}
\value{
A character vector prototype for schema definitions
}
\description{
Create a character type specification
}
\examples{
schema(name = chr())
}
