% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_results.R
\name{collect}
\alias{collect}
\alias{collect.parade_jobset}
\title{Collect results from all jobs in a jobset}
\usage{
collect(x, ...)

\method{collect}{parade_jobset}(x, simplify = TRUE, ...)
}
\arguments{
\item{x}{A \code{parade_jobset} object}

\item{...}{Additional arguments passed to method implementations}

\item{simplify}{Try to simplify results into a vector/matrix (default: TRUE)}
}
\value{
List or simplified structure of results
}
\description{
Collect results from all jobs in a jobset
}
\examples{
\dontrun{
jobs <- slurm_map(1:4, function(x) x^2)
await(jobs)
results <- collect(jobs)
}
}
