% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_class.R
\name{collect_result}
\alias{collect_result}
\title{Collect results from a job}
\usage{
collect_result(x)
}
\arguments{
\item{x}{A parade job object (parade_local_job or parade_script_job)}
}
\value{
The result value from the completed job.
}
\description{
Collect results from a job
}
\examples{
\dontrun{
job <- submit_slurm(my_fun)
await(job)
result <- collect_result(job)
}
}
