% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{cpus}
\alias{cpus}
\title{Set CPU count for a resource profile}
\usage{
cpus(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Number of CPUs}
}
\value{
Updated profile object
}
\description{
Set CPU count for a resource profile
}
\note{
Use res_cpus() to avoid naming collisions in user code.
}
\examples{
\donttest{
resources <- cpus(profile(), 16)
}

}
