% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_helpers.R
\name{digest}
\alias{digest}
\title{Generate job names from content digest}
\usage{
digest(prefix = "job", length = 8)
}
\arguments{
\item{prefix}{Prefix for the job name}

\item{length}{Number of characters from the hash to use}
}
\value{
A function suitable for use with \code{.name_by} parameter
}
\description{
Creates a naming function that uses a hash of the input for unique job names.
}
\note{
This function masks \code{digest::digest} when parade is attached. Prefer
\code{name_digest()} as an explicit alias to avoid confusion.
}
\examples{
\donttest{
if (interactive()) {
  # Each unique input gets a unique name
  data <- list(a = 1:10, b = 11:20, a = 1:10)  # Note: 'a' appears twice
  process_data <- function(x) x  # stub for example
  jobs <- slurm_map(data, process_data, .name_by = digest(), .engine = "local")
  # Job names will use hash, with identical inputs getting same hash
}
}

}
