% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_stage.R
\name{.guess_interpreter}
\alias{.guess_interpreter}
\title{Guess interpreter from script file extension}
\usage{
.guess_interpreter(script)
}
\arguments{
\item{script}{Path to a script file.}
}
\value{
A character scalar with the interpreter command.
}
\description{
Guess interpreter from script file extension
}
\keyword{internal}
