% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_log.R
\name{.log_append}
\alias{.log_append}
\title{Append lines to a log file}
\usage{
.log_append(path, lines)
}
\value{
NULL, called for side effect of writing to the log file.
}
\description{
Append lines to a log file
}
\keyword{internal}
