% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_log.R
\name{.pipeline_is_done}
\alias{.pipeline_is_done}
\title{Check if a deferred pipeline is done}
\usage{
.pipeline_is_done(d)
}
\value{
Logical scalar; TRUE if all chunks are done.
}
\description{
Check if a deferred pipeline is done
}
\keyword{internal}
