% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_log.R
\name{.pipeline_log_errors}
\alias{.pipeline_log_errors}
\title{Append error lines to the log}
\usage{
.pipeline_log_errors(errors_tbl, path)
}
\description{
Append error lines to the log
}
\keyword{internal}
