% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_log.R
\name{.pipeline_log_header}
\alias{.pipeline_log_header}
\title{Write the run header line to the log}
\usage{
.pipeline_log_header(d, path)
}
\description{
Write the run header line to the log
}
\keyword{internal}
