% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_log.R
\name{.pipeline_log_summary}
\alias{.pipeline_log_summary}
\title{Append the DONE summary line to the log}
\usage{
.pipeline_log_summary(d, path)
}
\value{
NULL, called for side effect of writing the summary.
}
\description{
Append the DONE summary line to the log
}
\keyword{internal}
