% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink_registry.R
\name{.register_builtin_formats}
\alias{.register_builtin_formats}
\title{Register built-in sink formats}
\usage{
.register_builtin_formats()
}
\description{
Register built-in sink formats
}
\keyword{internal}
