% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_parsers.R
\name{.slurm_exec}
\alias{.slurm_exec}
\title{Execution hook for testing}
\usage{
.slurm_exec(cmd, args)
}
\arguments{
\item{cmd}{Command to execute}

\item{args}{Command arguments}
}
\value{
Command output
}
\description{
Execution hook for testing
}
\keyword{internal}
