% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_parsers.R
\name{.slurm_sacct_info_v2}
\alias{.slurm_sacct_info_v2}
\title{Refactored sacct info with injectable executor}
\usage{
.slurm_sacct_info_v2(job_id, exec = .slurm_exec)
}
\description{
Refactored sacct info with injectable executor
}
\keyword{internal}
