% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{dry_run.default}
\alias{dry_run.default}
\title{Dry run for job submission (default)}
\usage{
\method{dry_run}{default}(
  .f,
  ...,
  .name = NULL,
  .resources = NULL,
  .write_result = NULL,
  .engine = "slurm"
)
}
\description{
Dry run for job submission (default)
}
