% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergonomics.R
\name{dry_run.parade_flow}
\alias{dry_run.parade_flow}
\title{Dry-run a flow: show plan and counts without executing}
\usage{
\method{dry_run}{parade_flow}(fl, ...)
}
\arguments{
\item{fl}{A \code{\link[=flow]{flow()}}.}
}
\description{
Dry-run a flow: show plan and counts without executing
}
