% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{explain.default}
\alias{explain.default}
\title{Explain a job submission (default)}
\usage{
\method{explain}{default}(
  .f,
  ...,
  .resources = NULL,
  .packages = character(),
  .engine = c("slurm", "local")
)
}
\description{
Explain a job submission (default)
}
