% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{file_ref}
\alias{file_ref}
\title{Create a file reference type specification}
\usage{
file_ref()
}
\value{
A packed schema for file reference metadata
}
\description{
Create a file reference type specification
}
\examples{
file_schema <- returns(output = file_ref())
}
