% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slurm.R
\name{find_file_arg}
\alias{find_file_arg}
\title{Find file-like argument in a list}
\usage{
find_file_arg(args)
}
\arguments{
\item{args}{List of arguments to search through}
}
\value{
First argument that looks like a file path, or NULL if none found
}
\description{
Find file-like argument in a list
}
\keyword{internal}
