% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_policy.R
\name{get_errors}
\alias{get_errors}
\title{Get collected errors from a policy}
\usage{
get_errors(policy)
}
\arguments{
\item{policy}{Error policy object}
}
\value{
List of collected errors
}
\description{
Get collected errors from a policy
}
\examples{
policy <- on_error("continue", collect_errors = TRUE)
get_errors(policy)
}
