% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waves.R
\name{get_running}
\alias{get_running}
\title{Get currently running jobs from a list}
\usage{
get_running(jobs)
}
\arguments{
\item{jobs}{List of job objects}
}
\value{
List of running jobs
}
\description{
Get currently running jobs from a list
}
\keyword{internal}
