% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{gpus}
\alias{gpus}
\title{Set GPU count for a resource profile}
\usage{
gpus(profile, value, type = NULL)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Number of GPUs}

\item{type}{Optional GPU type constraint}
}
\value{
Updated profile object
}
\description{
Set GPU count for a resource profile
}
\examples{
\donttest{
resources <- profile() \%>\% gpus(2)
resources <- profile() \%>\% gpus(1, type = "v100")
}

}
