% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{int}
\alias{int}
\title{Create an integer type specification}
\usage{
int()
}
\value{
An integer vector prototype for schema definitions
}
\description{
Create an integer type specification
}
\examples{
schema(count = int())
}
