% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-flex.R
\name{is_flex_type}
\alias{is_flex_type}
\title{Check if object is a flexible type specification}
\usage{
is_flex_type(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a flexible type spec
}
\description{
Check if object is a flexible type specification
}
\keyword{internal}
