% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waves.R
\name{is_flow_control}
\alias{is_flow_control}
\title{Check if object is a flow control policy}
\usage{
is_flow_control(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a flow control policy
}
\description{
Check if object is a flow control policy
}
