% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_class.R
\name{job_status}
\alias{job_status}
\title{Get job status}
\usage{
job_status(x)
}
\arguments{
\item{x}{A parade job object (parade_job, parade_script_job, or parade_local_job)}
}
\value{
A tibble containing job status information
}
\description{
Get job status
}
\examples{
\dontrun{
job <- submit_slurm(my_fun)
job_status(job)
}
}
