% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_results.R
\name{length.parade_jobset}
\alias{length.parade_jobset}
\title{Get length of jobset}
\usage{
\method{length}{parade_jobset}(x)
}
\arguments{
\item{x}{A parade_jobset object}
}
\value{
Integer; the number of jobs in the set.
}
\description{
Get length of jobset
}
\examples{
\dontrun{
jobs <- slurm_map(1:4, function(x) x^2)
length(jobs)
}

}
