% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{lgl}
\alias{lgl}
\title{Create a logical type specification}
\usage{
lgl()
}
\value{
A logical vector prototype for schema definitions
}
\description{
Create a logical type specification
}
\examples{
schema(is_valid = lgl())
}
