% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{lhs_grid}
\alias{lhs_grid}
\title{Create a Latin hypercube sample for parameter exploration}
\usage{
lhs_grid(n, ...)
}
\arguments{
\item{n}{Number of samples}

\item{...}{Named arguments with min/max ranges}
}
\value{
Data frame with sampled parameters
}
\description{
Generate a Latin hypercube sample for more efficient parameter
space exploration compared to regular grids.
}
\examples{
\donttest{
# Sample 20 points from parameter space
samples <- lhs_grid(
  n = 20,
  alpha = c(0, 1),      # min, max
  beta = c(0, 10),
  gamma = c(-1, 1)
)
}

}
