% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_backends.R
\name{list_submit_backends}
\alias{list_submit_backends}
\title{List available submit backends}
\usage{
list_submit_backends()
}
\value{
Character vector of backend names.
}
\description{
List available submit backends
}
