% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{lst}
\alias{lst}
\title{Create a list type specification}
\usage{
lst(ptype = list())
}
\arguments{
\item{ptype}{Prototype for list elements}
}
\value{
A list_of vector prototype for schema definitions
}
\description{
Create a list type specification
}
\examples{
schema(items = lst())
}
