% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{mem}
\alias{mem}
\title{Set memory limit for a resource profile}
\usage{
mem(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Memory limit (e.g., "16G", "32000M")}
}
\value{
Updated profile object
}
\description{
Set memory limit for a resource profile
}
\note{
Use res_mem() to avoid naming collisions in user code.
}
\examples{
\donttest{
resources <- mem(profile(), "32G")
}

}
