% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_helpers.R
\name{name_digest}
\alias{name_digest}
\title{Alias for digest-based naming that avoids masking \code{digest::digest}}
\usage{
name_digest(prefix = "job", length = 8)
}
\arguments{
\item{prefix}{Prefix for the job name}

\item{length}{Number of characters from the hash to use}
}
\value{
A naming function
}
\description{
Creates a naming function that uses a hash of the input for unique job names.
}
\examples{
name_digest("my_job", list(x = 1, y = "a"))
}
