% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{open_logs.default}
\alias{open_logs.default}
\title{Fallback log opener for list-like job objects}
\usage{
\method{open_logs}{default}(job, which = c("both", "out", "err"), viewer = utils::file.edit, ...)
}
\description{
Fallback log opener for list-like job objects
}
