% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_map.R
\name{packed_worker_function}
\alias{packed_worker_function}
\title{Worker function for packed execution}
\usage{
packed_worker_function(
  chunk_elements,
  chunk_indices,
  worker_fn,
  worker_args = list(),
  worker_extra_args = NULL,
  name_by = "auto",
  write_result_template = NULL,
  workers = 1L,
  parallel_backend = c("callr", "auto", "multicore", "multisession")
)
}
\description{
Processes a chunk of elements in parallel on a single node
}
\keyword{internal}
