% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_jobs.R
\name{parade_run_script_bt}
\alias{parade_run_script_bt}
\title{Run script in batchtools context}
\usage{
parade_run_script_bt(i, script, args, env, lib_paths, rscript, wd)
}
\arguments{
\item{i}{Job index}

\item{script}{Path to R script}

\item{args}{Command line arguments}

\item{env}{Environment variables}

\item{lib_paths}{Library paths to prepend}

\item{rscript}{Path to Rscript executable}

\item{wd}{Working directory}
}
\value{
List with ok status and exit code
}
\description{
Run script in batchtools context
}
\keyword{internal}
