% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{partition}
\alias{partition}
\title{Set partition for a resource profile}
\usage{
partition(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Partition name}
}
\value{
Updated profile object
}
\description{
Set partition for a resource profile
}
\note{
Use res_partition() to avoid naming collisions in user code.
}
