% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths_get}
\alias{paths_get}
\title{Get current parade path configuration}
\usage{
paths_get()
}
\value{
Named list of configured paths
}
\description{
Get current parade path configuration
}
\examples{
paths <- paths_get()
paths$data
}
