% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_filters.R
\name{pending}
\alias{pending}
\title{Select pending jobs}
\usage{
pending(x)
}
\arguments{
\item{x}{A \code{parade_jobset} object}
}
\value{
A parade_jobset containing only pending jobs
}
\description{
Select pending jobs
}
