% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_policy.R
\name{print.parade_error_policy}
\alias{print.parade_error_policy}
\title{Print method for error policies}
\usage{
\method{print}{parade_error_policy}(x, ...)
}
\arguments{
\item{x}{Error policy object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for error policies
}
\examples{
policy <- on_error("continue", collect_errors = TRUE)
print(policy)

}
