% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waves.R
\name{print.parade_flow_control}
\alias{print.parade_flow_control}
\title{Print method for flow control policies}
\usage{
\method{print}{parade_flow_control}(x, ...)
}
\arguments{
\item{x}{Flow control policy}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for flow control policies
}
