% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_class.R
\name{print.parade_local_job}
\alias{print.parade_local_job}
\title{Print method for local jobs}
\usage{
\method{print}{parade_local_job}(x, ...)
}
\arguments{
\item{x}{A parade_local_job object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the job object
}
\description{
Print method for local jobs
}
