% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_jobs.R
\name{print.parade_script_job}
\alias{print.parade_script_job}
\title{Print method for parade script jobs}
\usage{
\method{print}{parade_script_job}(x, ...)
}
\arguments{
\item{x}{A \code{parade_script_job} object}

\item{...}{Additional arguments (ignored)}
}
\value{
The input object (invisibly)
}
\description{
Print method for parade script jobs
}
\examples{
\dontrun{
job <- submit_slurm("analysis.R", resources = list(time = "1:00:00"))
print(job)
}

}
