% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{res_account}
\alias{res_account}
\title{Alias for account()}
\usage{
res_account(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Account name}
}
\value{
Updated profile
}
\description{
Alias for account()
}
