% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{res_cpus}
\alias{res_cpus}
\title{Alias for cpus()}
\usage{
res_cpus(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Number of CPUs}
}
\value{
Updated profile
}
\description{
Alias for cpus()
}
\examples{
res_cpus(profile(), 4)
}
