% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{res_partition}
\alias{res_partition}
\title{Alias for partition()}
\usage{
res_partition(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Partition name}
}
\value{
Updated profile
}
\description{
Alias for partition()
}
\examples{
res_partition(profile(), "gpu")
}
