% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{res_time}
\alias{res_time}
\title{Set time limit for a resource profile (non-masking)}
\usage{
res_time(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Time limit (e.g., "4:00:00", "2-00:00:00")}
}
\value{
Updated profile
}
\description{
Use \code{res_time()} in pipelines to avoid masking \code{stats::time()} when \code{parade} is attached.
}
