% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_policy.R
\name{resubmit_job}
\alias{resubmit_job}
\title{Resubmit a failed job}
\usage{
resubmit_job(job)
}
\arguments{
\item{job}{Job object to resubmit}
}
\value{
New job object
}
\description{
Resubmit a failed job
}
\keyword{internal}
