% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_job.R
\name{scaffold_batch_template}
\alias{scaffold_batch_template}
\title{Create a basic SLURM batch template}
\usage{
scaffold_batch_template(
  system = c("slurm"),
  out = file.path("batchtools", paste0("parade-", match.arg(system), ".tmpl")),
  modules = "R",
  exports = c(PARADE_SCRATCH = "${SLURM_TMPDIR:-${TMPDIR:-/tmp}}/parade-$SLURM_JOB_ID",
    OMP_NUM_THREADS = "1", MKL_NUM_THREADS = "1", OPENBLAS_NUM_THREADS = "1"),
  preamble = character(),
  overwrite = FALSE
)
}
\arguments{
\item{path}{Path where template should be created (temp file if NULL)}
}
\value{
Path to created template file (invisibly)
}
\description{
Create a basic SLURM batch template
}
\examples{
template_path <- scaffold_batch_template()
}
