% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_job.R
\name{scaffold_flow_job}
\alias{scaffold_flow_job}
\title{Generate scaffold scripts for SLURM flow execution}
\usage{
scaffold_flow_job(
  flow,
  name = "parade_job",
  registry_dir = NULL,
  dir = getwd(),
  modules = NULL,
  exports = NULL
)
}
\arguments{
\item{flow}{A \code{parade_flow} object}

\item{name}{Base name for generated scripts}

\item{registry_dir}{Registry directory for job execution}

\item{dir}{Directory where scripts should be created}

\item{modules}{SLURM modules to load}

\item{exports}{Environment variables to export}
}
\value{
List of created script paths (invisibly)
}
\description{
Creates a set of helper scripts for submitting, monitoring, and collecting
results from a parade flow on SLURM systems.
}
\examples{
\donttest{
flow <- flow(data.frame(x = 1:3))
scaffold_flow_job(flow, name = "my_job", dir = tempdir())
}
}
