% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_jobs.R
\name{script_cancel}
\alias{script_cancel}
\title{Cancel a running SLURM script job}
\usage{
script_cancel(job)
}
\arguments{
\item{job}{A \code{parade_script_job} object}
}
\value{
The input job object (invisibly)
}
\description{
Cancel a running SLURM script job
}
\examples{
\donttest{
job <- submit_slurm("script.R")
script_cancel(job)
}
}
