% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_jobs.R
\name{script_load}
\alias{script_load}
\title{Load a script job from its registry directory}
\usage{
script_load(registry_dir)
}
\arguments{
\item{registry_dir}{Path to the batchtools registry directory}
}
\value{
A \code{parade_script_job} object
}
\description{
Load a script job from its registry directory
}
\examples{
\donttest{
job <- script_load("/path/to/registry")
}
}
