% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_stats.R
\name{script_logs}
\alias{script_logs}
\title{Get log file paths for a SLURM job}
\usage{
script_logs(job)
}
\arguments{
\item{job}{A \code{parade_script_job} object}
}
\value{
Tibble with log file paths and modification times
}
\description{
Get log file paths for a SLURM job
}
\examples{
\donttest{
if (Sys.which("squeue") != "") {
  job <- submit_slurm("script.R")
  logs <- script_logs(job)
}
}
}
