% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_jobs.R
\name{script_status}
\alias{script_status}
\title{Get status of a SLURM script job}
\usage{
script_status(job, detail = FALSE)
}
\arguments{
\item{job}{A \code{parade_script_job} object}

\item{detail}{Whether to return detailed job information}
}
\value{
A tibble with job status information
}
\description{
Get status of a SLURM script job
}
\examples{
\donttest{
job <- submit_slurm("script.R")
status <- script_status(job)
}
}
