% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_stats.R
\name{script_tail}
\alias{script_tail}
\title{Display recent log output from a SLURM job}
\usage{
script_tail(job, n = 200)
}
\arguments{
\item{job}{A \code{parade_script_job} object}

\item{n}{Number of lines to show from end of log}
}
\value{
Log lines (invisibly)
}
\description{
Display recent log output from a SLURM job
}
\examples{
\donttest{
if (Sys.which("squeue") != "") {
  job <- submit_slurm("script.R")
  script_tail(job, n = 50)
}
}
}
