% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_macros.R
\name{simple_expand}
\alias{simple_expand}
\title{Simple macro expansion without glue}
\usage{
simple_expand(template, replacements)
}
\arguments{
\item{template}{Template string}

\item{replacements}{Named list of replacements}
}
\value{
Expanded string
}
\description{
Simple macro expansion without glue
}
\keyword{internal}
