% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{slurm_template}
\alias{slurm_template}
\title{Get path to default SLURM template}
\usage{
slurm_template()
}
\value{
Path to package SLURM template file
}
\description{
Get path to default SLURM template
}
\examples{
template_path <- slurm_template()
}
