% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{slurm_template_default}
\alias{slurm_template_default}
\title{Get the default Slurm template from config or package}
\usage{
slurm_template_default()
}
\description{
Get the default Slurm template from config or package
}
