% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{status}
\alias{status}
\title{Get status of all jobs in a jobset}
\usage{
status(x, ...)
}
\arguments{
\item{x}{A parade_jobset object}
}
\value{
A tibble with job status information
}
\description{
Get status of all jobs in a jobset
}
