% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{struct}
\alias{struct}
\title{Alias for pack function}
\usage{
struct(.returns)
}
\arguments{
\item{.returns}{Schema specification}
}
\value{
A \code{parade_pack} object
}
\description{
Alias for pack function
}
\examples{
struct(returns(name = chr(), value = dbl()))
}
