% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_results.R
\name{[.parade_jobset}
\alias{[.parade_jobset}
\title{Extract subset of jobs}
\usage{
\method{[}{parade_jobset}(x, i)
}
\arguments{
\item{x}{A parade_jobset object}

\item{i}{Index vector for subsetting}
}
\value{
A \code{parade_jobset} containing the selected jobs.
}
\description{
Extract subset of jobs
}
\examples{
\dontrun{
jobs <- slurm_map(1:4, function(x) x^2)
first_two <- jobs[1:2]
}

}
