% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_results.R
\name{tail.parade_jobset}
\alias{tail.parade_jobset}
\title{Get tail of logs for jobs}
\usage{
\method{tail}{parade_jobset}(x, n = 50, ...)
}
\arguments{
\item{x}{A parade_jobset or parade_job}

\item{n}{Number of lines to show}

\item{...}{Additional arguments}
}
\value{
Character vector of log lines (invisibly)
}
\description{
Get tail of logs for jobs
}
