% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{tbl}
\alias{tbl}
\title{Create a tibble/data frame type specification}
\usage{
tbl(...)
}
\arguments{
\item{...}{Column specifications for the tibble}
}
\value{
A tibble prototype for schema definitions
}
\description{
Create a tibble/data frame type specification
}
\examples{
schema(data = tbl())
}
