% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{time}
\alias{time}
\title{Set time limit for a resource profile}
\usage{
time(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Time limit (e.g., "4:00:00", "2-00:00:00")}
}
\value{
Updated profile object
}
\description{
Set time limit for a resource profile
}
\note{
masks stats::time when parade is attached; use res_time() to avoid masking.
}
\examples{
\donttest{
resources <- res_time(profile(), "8:00:00")
}

}
