% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitoring.R
\name{use_registry}
\alias{use_registry}
\title{Get or set active registry}
\usage{
use_registry(path = NULL, create = TRUE)
}
\arguments{
\item{path}{Path to registry directory}

\item{create}{If TRUE, create directory if it doesn't exist}
}
\value{
Current registry path
}
\description{
Manage which registry is currently active for job storage.
}
\examples{
\donttest{
# Get current registry
use_registry()

# Switch to different registry
use_registry("~/my_project/jobs")
}

}
