% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergonomics.R
\name{with_parade_options}
\alias{with_parade_options}
\title{Temporarily set parade options for code execution}
\usage{
with_parade_options(..., code)
}
\arguments{
\item{...}{Named parade option values to set temporarily}

\item{code}{Code to execute with modified options}
}
\value{
Result of executing the code
}
\description{
Temporarily set parade options for code execution
}
\examples{
with_parade_options(error = "stop", code = {
  message("running with error = 'stop'")
  1 + 1
})
}
