% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{as_tibble.parade_jobset}
\alias{as_tibble.parade_jobset}
\title{Convert jobset to tibble}
\usage{
\method{as_tibble}{parade_jobset}(x, ...)
}
\arguments{
\item{x}{A parade_jobset object}

\item{...}{Additional arguments (unused)}
}
\description{
Convert jobset to tibble
}
