% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_await.R
\name{await}
\alias{await}
\alias{await.parade_jobset}
\title{Wait for jobs to complete}
\usage{
await(x, ...)

\method{await}{parade_jobset}(x, timeout = Inf, poll = 10, .progress = NULL, ...)
}
\arguments{
\item{x}{A \code{parade_jobset} or parade job object}

\item{...}{Additional arguments passed to method implementations}

\item{timeout}{Maximum time to wait in seconds}

\item{poll}{Polling interval in seconds}

\item{.progress}{Show progress bar}
}
\value{
The jobset/job (invisibly)
}
\description{
Wait for jobs to complete
}
\section{Methods (by class)}{
\itemize{
\item \code{await(parade_jobset)}: Wait for all jobs in a jobset to complete

}}
\examples{
\dontrun{
jobs <- slurm_map(1:4, function(x) x^2)
await(jobs, timeout = 600)
}
}
