% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{await}
\alias{await}
\title{Wait for all jobs in a jobset to complete}
\usage{
await(x, ...)
}
\arguments{
\item{x}{A parade_jobset object}

\item{timeout}{Maximum time to wait in seconds}

\item{poll}{Polling interval in seconds}

\item{.progress}{Show progress bar}
}
\value{
The jobset (invisibly)
}
\description{
Wait for all jobs in a jobset to complete
}
