% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_results.R
\name{cancel}
\alias{cancel}
\title{Cancel all jobs in a jobset}
\usage{
cancel(x, ...)
}
\arguments{
\item{x}{A \code{parade_jobset} object}

\item{...}{Additional arguments passed to method implementations}
}
\value{
The jobset (invisibly)
}
\description{
Cancel all jobs in a jobset
}
\examples{
\dontrun{
jobs <- slurm_map(1:4, function(x) Sys.sleep(100))
cancel(jobs)
}
}
