% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{collect}
\alias{collect}
\title{Collect results from all jobs in a jobset}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{A parade_jobset object}

\item{simplify}{Try to simplify results into a vector/matrix}
}
\value{
List or simplified structure of results
}
\description{
Collect results from all jobs in a jobset
}
