% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_filters.R
\name{completed}
\alias{completed}
\title{Select completed jobs}
\usage{
completed(x)
}
\arguments{
\item{x}{A \code{parade_jobset} object}
}
\value{
A parade_jobset containing only completed jobs
}
\description{
Select completed jobs
}
