% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deferred.R
\name{deferred_cancel}
\alias{deferred_cancel}
\title{Cancel deferred execution jobs}
\usage{
deferred_cancel(d, which = c("running", "all"))
}
\arguments{
\item{d}{A \code{parade_deferred} object}

\item{which}{Which jobs to cancel: "running" or "all"}
}
\value{
The input deferred object (invisibly)
}
\description{
Cancel deferred execution jobs
}
\examples{
\donttest{
deferred <- submit(fl)
deferred_cancel(deferred, which = "running")
}
}
