% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_parsers.R
\name{.slurm_squeue_info_v2}
\alias{.slurm_squeue_info_v2}
\title{Refactored squeue info with injectable executor}
\usage{
.slurm_squeue_info_v2(job_id, exec = .slurm_exec)
}
\arguments{
\item{job_id}{Job ID to query}

\item{exec}{Execution function (defaults to .slurm_exec)}
}
\value{
Parsed squeue output
}
\description{
Refactored squeue info with injectable executor
}
\keyword{internal}
