% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergonomics.R
\name{explain.parade_flow}
\alias{explain.parade_flow}
\title{Explain a flow: DAG + distribution + sinks}
\usage{
\method{explain}{parade_flow}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=flow]{flow()}}.}

\item{...}{Additional arguments passed to methods (unused).}
}
\value{
A tibble summarizing stages.
}
\description{
Explain a flow: DAG + distribution + sinks
}
