% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergonomics.R
\name{explain}
\alias{explain}
\title{Explain a flow: DAG + distribution + sinks}
\usage{
explain(fl)
}
\arguments{
\item{fl}{A \code{\link[=flow]{flow()}}.}
}
\value{
A tibble summarizing stages.
}
\description{
Explain a flow: DAG + distribution + sinks
}
