% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{failed}
\alias{failed}
\title{Select failed jobs}
\usage{
failed(x, stage = NULL, ...)
}
\arguments{
\item{x}{A parade_jobset object}

\item{stage}{Optional stage filter (ignored for jobsets)}

\item{...}{Additional arguments (ignored)}
}
\description{
Select failed jobs
}
