% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_class.R
\name{is_done}
\alias{is_done}
\title{Check if a job is done}
\usage{
is_done(x)
}
\arguments{
\item{x}{A parade job object (parade_job, parade_script_job, or parade_local_job)}
}
\value{
Logical scalar; TRUE if the job has completed.
}
\description{
Check if a job is done
}
\examples{
\dontrun{
job <- submit_slurm(my_fun)
is_done(job)
}
}
