% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{param_grid}
\alias{param_grid}
\title{Create a parameter grid using tidyr::crossing}
\usage{
param_grid(...)
}
\arguments{
\item{...}{Named vectors or lists to cross}
}
\value{
A tibble with all parameter combinations
}
\description{
Create a parameter grid using tidyr::crossing
}
\examples{
grid <- param_grid(x = 1:3, method = c("A", "B"))
}
