% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_class.R
\name{print.parade_job}
\alias{print.parade_job}
\title{Print method for parade job objects}
\usage{
\method{print}{parade_job}(x, ...)
}
\arguments{
\item{x}{A parade_job object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the job object
}
\description{
Common interface for printing all parade job types (script, function, local)
}
\examples{
\dontrun{
job <- submit_slurm("my_script.R", resources = list(time = "1:00:00"))
print(job)
}

}
