% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_as_jobset.R
\name{print.parade_jobset}
\alias{print.parade_jobset}
\title{Print method for parade_jobset}
\usage{
\method{print}{parade_jobset}(x, ...)
}
\arguments{
\item{x}{A parade_jobset object to print}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for parade_jobset
}
