% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{profile_clear}
\alias{profile_clear}
\title{Clear all registered profiles}
\usage{
profile_clear()
}
\value{
Invisible NULL
}
\description{
Clear all registered profiles
}
\examples{
\dontrun{
profile_clear()
}
}
