% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{profile_init_defaults}
\alias{profile_init_defaults}
\title{Initialize default resource profiles}
\usage{
profile_init_defaults(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Whether to overwrite existing profiles}
}
\value{
Invisible NULL
}
\description{
Set up commonly used resource profiles. This function is called
automatically when the package is loaded but can be called manually
to reset profiles.
}
\examples{
\donttest{
# Reset to default profiles
profile_init_defaults(overwrite = TRUE)
}

}
