% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_profiles.R
\name{res_time}
\alias{res_time}
\title{Alias for time() that avoids masking stats::time}
\usage{
res_time(profile, value)
}
\arguments{
\item{profile}{A resource profile object}

\item{value}{Time limit (e.g., "4:00:00", "2-00:00:00")}
}
\value{
Updated profile
}
\description{
Alias for time() that avoids masking stats::time
}
