% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{running}
\alias{running}
\title{Select running jobs}
\usage{
running(x)
}
\arguments{
\item{x}{A parade_jobset object}
}
\value{
A parade_jobset containing only running jobs
}
\description{
Select running jobs
}
