% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{slurm_defaults_get}
\alias{slurm_defaults_get}
\title{Get defaults for SLURM (merged from options() and config)}
\usage{
slurm_defaults_get(profile = "default")
}
\arguments{
\item{profile}{character name; 'default' by default}
}
\description{
Get defaults for SLURM (merged from options() and config)
}
