% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{slurm_defaults_set}
\alias{slurm_defaults_set}
\title{Set defaults for SLURM (R session and optionally persist to config)}
\usage{
slurm_defaults_set(..., .list = NULL, profile = "default", persist = FALSE)
}
\arguments{
\item{...}{key=value pairs (e.g., mem = NA, time = "2h")}

\item{.list}{optional named list}

\item{profile}{profile name; defaults to 'default'}

\item{persist}{write to config file if TRUE}
}
\description{
Set defaults for SLURM (R session and optionally persist to config)
}
