% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{slurm_resources}
\alias{slurm_resources}
\title{Merge resources with defaults + normalize + drop NA/omit}
\usage{
slurm_resources(resources = NULL, profile = "default")
}
\description{
Merge resources with defaults + normalize + drop NA/omit
}
