% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{slurm_template_default}
\alias{slurm_template_default}
\title{Get the default SLURM template path}
\usage{
slurm_template_default()
}
\value{
Path to SLURM template file
}
\description{
Retrieves the configured SLURM template path or falls back to the
package default template.
}
\examples{
template_path <- slurm_template_default()
}
