% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobset_class.R
\name{[.parade_jobset}
\alias{[.parade_jobset}
\title{Extract subset of jobs}
\usage{
\method{[}{parade_jobset}(x, i)
}
\arguments{
\item{x}{A parade_jobset object}

\item{i}{Index vector for subsetting}
}
\description{
Extract subset of jobs
}
