% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_jobs.R
\name{submit_slurm}
\alias{submit_slurm}
\title{Submit an R script to SLURM and monitor it from R}
\usage{
submit_slurm(
  script,
  args = character(),
  name = NULL,
  template = NULL,
  resources = NULL,
  registry_dir = NULL,
  env = character(),
  lib_paths = .libPaths(),
  rscript = file.path(R.home("bin"), "Rscript"),
  wd = dirname(normalizePath(script))
)
}
\description{
Submit an R script to SLURM and monitor it from R
}
